pyproject.toml

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
import_heading_stdlib = "Standard Library"
import_heading_firstparty = "Our Libraries"
import_heading_thirdparty = "Third Party Libraries"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --color=yes"


[tool.poetry]
name = "strava-mongo-lambda"
version = "1.0.0"
description = "Function URL Lambda with FastAPI docker image secured by Cognito"
authors = ["Josh Peak <neozenith.dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pymongo = {extras = ["srv"], version = "^4.3.2"}
python-dotenv = "^0.21.0"
swagger-client = {path = "strava.zip"}
fastapi = {extras = ["all"], version = "^0.86.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
jinja2 = "3.0.3"
gspread = "^5.6.2"
httpx = "^0.23.0"
mangum = "^0.16.0"

[tool.poetry.group.dev.dependencies]
invoke-common-tasks = {extras = ["all"], version = "^0.3.2"}
invoke = "^1.7.3"
boto3 = "^1.26.4"


[tool.mypy]
exclude = ["tests/", "tasks\\.py"]
pretty = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
ignore_missing_imports = true
follow_imports = "silent"
disallow_incomplete_defs = true
disallow_untyped_defs = false
strict = false

[tool.coverage]
[tool.coverage.run]
branch = true
omit = ["tests/*", "**/__init__.py", "tasks.py"]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
